refactor(mt#797): split rules-command-operations.ts and taskCommands.ts into focused sub-modules
rules-command-operations.ts (729 lines) → operations/types.ts, operations/config-operations.ts,
operations/crud-operations.ts, operations/migration-operations.ts; barrel is now ~55 lines.

taskCommands.ts (695 lines) → commands/shared-helpers.ts, commands/query-commands.ts,
commands/mutation-commands.ts; barrel is now ~30 lines.

Also added commands/shared-helpers.ts to ESLint no-singleton-reach-in allowedFiles since it
carries the getSharedSessionProvider call that was previously allowed in taskCommands.ts.

All original public exports preserved.
